Module Tezos_crypto__Pvss

PVSS protocol, following

see Schoenmakers, B., 1999: A simple publicly verifiable secret sharing scheme and its application to electronic voting. Lecture Notes in Computer Science, pp.148-164.

see https://www.win.tue.nl/~berry/papers/crypto99.pdf

The protocol is expressed as a functor parametrized by a cyclic group of prime order. Algebraic properties are enforced at the type level, whenever reasonably possible.

module type CYCLIC_GROUP = sig ... end

module type PVSS = sig ... end

PVSS construction, based on a cyclic group G of prime order

module MakePvss : functor (G : CYCLIC_GROUP) -> PVSS